requirements.txt
msal==1.8.0
msal-extensions==0.3.0
oauthlib==3.1.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.3
# waitress==1.4.4
pytz==2021.1
python-configuration==0.8.2
python-dateutil==2.8.1
